more unified device cleanup
